requirements.txt
aiogram~=2.25.1
python-dotenv~=1.0.0
fastapi~=0.95.2
openai~=1.9.0